Moved from SUI dropdown hotfix from the dropdown component to the select component
